Add compact variant of govuk-accordion
There is risk to this approach, where a custom component
is more proper. For now, since overrides already existed,
adjust the overrides to better render with the changes
to the accordion which were introduced in v4.

In future if this page is redesigned consider creating a new filter component
similar to GOV.UK's Search page.